Refactor commands into commands.rs
